implemented many to many relationship with posts and tags, fixed bug allowing tags and categories to be deleted while still associated with a post, implemented admins, and user profiles

+++ app/Http/Controllers/CategoriesController.php
@@ -90,7 +90,10 @@ class CategoriesController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy(Category $category) {
-
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category cannot be deleted because it is associated with a post');
+            return redirect()->back();
+        }
         $category->delete(); 
 
         session()->flash('success', 'Category deleted successfully.');
